Support decoding optional values hyperbee skips (en/de)coding (#26)
The `value` in `bee.put(key, value)` is optional and is not encoded via
the `valueEncoding` option when the node is appended. Because of this,
`hyperbee-diff-stream` threw an error trying to decode a `null`
assuming it was a buffer. A new check for `null` skips decoding allowing
it to pass through.